fix: routes name, change variable name to Camel case
