+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './bases'
+export * from './common'
