+++ src/screens/index.ts
@@ -4,5 +4,5 @@ export * from './VerificationCodeScreen'
 export * from './SplashScreen'
 export * from './NavigateScreen'
 export * from './PhoneRegistrationScreen'
-export * from './Login'
+export * from './LoginScreen'
 export * from './Register'
